Change structure
Start to use cargo workspaces, now it has two crates:
gui - user interface
y_project - lib with cli
