+++ gui/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "gui"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+relm4 = "0.9"
+relm4-components = "0.9"
