--- src/util.rs
@@ -1,5 +1,0 @@
-pub mod elf_structure;
-pub mod jvm_structure;
-pub mod macho_structure;
-pub mod pe_structure;
-pub mod signature;
