+++ y_project/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "y_project"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+name = "y_project"
+path = "src/lib.rs"
+
+[[bin]]
+name="main"
+path = "src/main.rs"
+
+[dependencies]
