+++ y_project/src/main.rs
@@ -1,0 +1,9 @@
+use y_project::*;
+
+fn main() {
+    help();
+    //let context: Ctx = get_arguments();
+    //let bytecode = read_file(&context.filename);
+    //let sign = get_sign(&bytecode);
+    //get_file_data(&sign, &bytecode);
+}
